refactor: move mapStatementsToPathInZip from SegmentsByStatementsExporter to FileNameGenerator
Censoring is now pre-computed per statement in the controller using
needsToBeCensored, and passed as [Statement, bool] pairs to
FileNameGenerator::mapStatementsToPathInZip. This keeps censoring logic
in the export domain and FileNameGenerator focused on naming concerns.

Moves related tests from SegmentsByStatementsExporterTest to FileNameGeneratorTest.
