+++ tests/backend/core/Statement/Functional/FileNameGeneratorTest.php
@@ -17,6 +17,9 @@ use demosplan\DemosPlanCoreBundle\DataGenerator\Factory\Statement\StatementFacto
 use demosplan\DemosPlanCoreBundle\DataGenerator\Factory\Statement\StatementMetaFactory;
 use demosplan\DemosPlanCoreBundle\Entity\Statement\Statement;
 use demosplan\DemosPlanCoreBundle\Entity\Statement\StatementMeta;
+use demosplan\DemosPlanCoreBundle\Entity\User\Orga;
+use demosplan\DemosPlanCoreBundle\Entity\User\User;
+use demosplan\DemosPlanCoreBundle\Exception\InvalidArgumentException;
 use demosplan\DemosPlanCoreBundle\Logic\Segment\Export\FileNameGenerator;
 use Tests\Base\FunctionalTestCase;
 use Zenstruck\Foundry\Persistence\Proxy;
@@ -62,6 +65,128 @@ class FileNameGeneratorTest extends FunctionalTestCase
                 $templateName,
                 $this->testStatement);
         }
+    }
+
+    public function testMapStatementsToPathInZipWithTrueDuplicate(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->sut->mapStatementsToPathInZip([
+            [$this->testStatement->_real(), false],
+            [$this->testStatement->_real(), false],
+        ], '');
+    }
+
+    /**
+     * @dataProvider getCensorParams
+     */
+    public function testCensorshipOnPathOnExportSegmentsInZip(
+        bool $censorCitizenData,
+        bool $censorInstitutionData,
+    ): void {
+        $citizenOrganisation = $this->find(Orga::class, User::ANONYMOUS_USER_ORGA_ID);
+
+        $internalStatement = StatementFactory::createOne();
+        $externalStatement = StatementFactory::createOne(['organisation' => $citizenOrganisation]);
+
+        static::assertTrue($externalStatement->isSubmittedByCitizen());
+        static::assertTrue($internalStatement->isSubmittedByOrganisation());
+
+        $statements = $this->sut->mapStatementsToPathInZip([
+            [$externalStatement->_real(), $externalStatement->isSubmittedByCitizen() && $censorCitizenData],
+            [$internalStatement->_real(), $internalStatement->isSubmittedByOrganisation() && $censorInstitutionData],
+        ]);
+
+        foreach ($statements as $key => $statement) {
+            if ($statement->isSubmittedByCitizen()) {
+                if ($censorCitizenData) {
+                    $expectedAKey = $statement->getExternId().'.docx';
+                    static::assertEquals($expectedAKey, $key);
+                } else {
+                    $expectedAKey = $statement->getExternId().'-einreichende-person-unbekannt-eingangsnummer-unbekannt.docx';
+                    static::assertEquals($expectedAKey, $key);
+                }
+            } elseif ($statement->isSubmittedByOrganisation()) {
+                if ($censorInstitutionData) {
+                    $expectedAKey = $statement->getExternId().'.docx';
+                    static::assertEquals($expectedAKey, $key);
+                } else {
+                    $expectedAKey = $statement->getExternId().'-einreichende-person-unbekannt-eingangsnummer-unbekannt.docx';
+                    static::assertEquals($expectedAKey, $key);
+                }
+            }
+        }
+    }
+
+    /**
+     * @dataProvider getCensorParams
+     */
+    public function testMapStatementsToPathInZipWithSuperficialDuplicate(
+        bool $censorCitizenData,
+        bool $censorInstitutionData,
+    ): void {
+        $statementA = $this->createMinimalTestStatement('a', 'a', 'a');
+        $statementB = $this->createMinimalTestStatement('b', 'a', 'a');
+
+        $censoredA = ($censorCitizenData && $statementA->isSubmittedByCitizen()) || ($censorInstitutionData && $statementA->isSubmittedByOrganisation());
+        $censoredB = ($censorCitizenData && $statementB->isSubmittedByCitizen()) || ($censorInstitutionData && $statementB->isSubmittedByOrganisation());
+
+        $statements = $this->sut->mapStatementsToPathInZip([
+            [$statementA->_real(), $censoredA],
+            [$statementB->_real(), $censoredB],
+        ]);
+
+        if ($censoredA) {
+            $expectedAKey = 'statement-extern-id-a.docx';
+        } else {
+            $expectedAKey = 'statement-extern-id-a-statement-author-name-a-statement-intern-id-a.docx';
+        }
+        self::assertArrayHasKey($expectedAKey, $statements);
+        self::assertSame($statementA->_real(), $statements[$expectedAKey]);
+
+        if ($censoredB) {
+            $expectedBKey = 'statement-extern-id-b.docx';
+        } else {
+            $expectedBKey = 'statement-extern-id-b-statement-author-name-a-statement-intern-id-a.docx';
+        }
+        self::assertArrayHasKey($expectedBKey, $statements);
+        self::assertSame($statementB->_real(), $statements[$expectedBKey]);
+    }
+
+    /**
+     * @dataProvider getCensorParams
+     */
+    public function testMapStatementsToPathInZipWithoutDuplicate(
+        bool $censorCitizenData,
+        bool $censorInstitutionData,
+    ): void {
+        $statementA = $this->createMinimalTestStatement('xyz', 'xyz', 'xyz');
+        $statementB = $this->createMinimalTestStatement('xyz', 'xyz', 'xyz');
+
+        $censoredA = ($censorCitizenData && $statementA->isSubmittedByCitizen()) || ($censorInstitutionData && $statementA->isSubmittedByOrganisation());
+        $censoredB = ($censorCitizenData && $statementB->isSubmittedByCitizen()) || ($censorInstitutionData && $statementB->isSubmittedByOrganisation());
+
+        $statements = $this->sut->mapStatementsToPathInZip([
+            [$statementA->_real(), $censoredA],
+            [$statementB->_real(), $censoredB],
+        ], '');
+
+        $expectedAKey = 'statement-extern-id-xyz-statement-author-name-xyz-statement-intern-id-xyz-'.$statementA->getId().'.docx';
+        self::assertArrayHasKey($expectedAKey, $statements);
+        self::assertSame($statementA->_real(), $statements[$expectedAKey]);
+
+        $expectedBKey = 'statement-extern-id-xyz-statement-author-name-xyz-statement-intern-id-xyz-'.$statementB->getId().'.docx';
+        self::assertArrayHasKey($expectedBKey, $statements);
+        self::assertSame($statementB->_real(), $statements[$expectedBKey]);
+    }
+
+    public function getCensorParams(): array
+    {
+        return [
+            [true, true],
+            [false, false],
+            [true, false],
+            [false, true],
+        ];
     }
 
     private function verifyFileNameFromTemplate(string $rawExpectedFileName, string $templateName, Statement|Proxy|null $testStatement): void
